Se cambia menu de reportes a la vista del seller
