+++ api/controllers/SellerController.js
@@ -20,6 +20,7 @@ module.exports = {
     let seller = null;
     let integrations = null;
     let commissiondiscount = null;
+    let months = [];
     let action = req.param('action') ? req.param('action') : null;
     let id = req.param('id') ? req.param('id') : null;
     if(rights.name!=='superadmin' && rights.name!=='admin'){
@@ -28,6 +29,8 @@ module.exports = {
       sellers = await Seller.find();
     }
     if(id){
+      let currentDay =  moment().format('DD');
+      let availableOptions = false;
       seller = await Seller.findOne({id:id}).populate('mainAddress');
       if(seller.mainAddress!==undefined && seller.mainAddress!==null){
         seller.mainAddress = await Address.findOne({id:seller.mainAddress.id})
@@ -41,9 +44,17 @@ module.exports = {
       commissiondiscount = await CommissionDiscount.find({
         where:{seller:id},
       });
+      for (let i = 14; i >= 0; i--) {
+        let month = moment().subtract(i+1, 'months').locale('es').format('MMMM YYYY');
+        let available = moment().subtract(i, 'months').locale('es').format('MMMM YYYY');
+        if (i === 0 && currentDay < 20) {
+          availableOptions = true;
+        }
+        months.push({month, available, availableOptions});
+      }
     }
     let countries = await Country.find();
-    res.view('pages/sellers/sellers',{layout:'layouts/admin',sellers:sellers,action:action,seller:seller,error:error,success:success,countries:countries, integrations, commissiondiscount, appIdMl: constant.APP_ID_ML, secretKeyMl: constant.SECRET_KEY_ML, moment});
+    res.view('pages/sellers/sellers',{layout:'layouts/admin',sellers:sellers,months,action:action,seller:seller,error:error,success:success,countries:countries, integrations, commissiondiscount, appIdMl: constant.APP_ID_ML, secretKeyMl: constant.SECRET_KEY_ML, moment});
   },
   createseller: async function(req, res){
     let rights = await sails.helpers.checkPermissions(req.session.user.profile);
