--- views/pages/sellers/reports.ejs
@@ -1,138 +1,0 @@
-<script src="https://cdnjs.cloudflare.com/ajax/libs/FileSaver.js/2.0.0/FileSaver.min.js"></script>
-<section class="container">
-  <h4 class="title is-4">Informes Mensuales de Ventas</h4>
-  <hr>
-  <div id="loading-block" class="notification is-hidden">
-    <i class="fas fa-sync fa-spin"></i>
-      Estamos generando tu informe...
-      <progress class="progress is-small is-primary" max="100">15%</progress>
-  </div>
-  <div class="box content">
-    <%if(sellers && sellers.length>0){%>
-      <div class="field">
-        <label class="label">Elegir Seller</label>
-        <div class="select is-fullwidth">
-          <select name="seller" id="selectSeller" required>
-            <option class="is-capitalized" value="">-- Elegir Seller --</option>
-            <% sellers.forEach(s=>{%> {%>
-              <option value="<%=s.id%>"><%=s.name%></option>
-            <% }) %>
-          </select>
-        </div>
-      </div>
-    <% } %>
-    <br>
-    <table id="reportTable" class="table is-striped is-hidden">
-      <thead>
-        <tr>
-          <th>Mes</th>
-          <th>Fecha Disponibilidad</th>
-          <th></th>
-          <th></th>
-          <th></th>
-        </tr>
-      </thead>
-      <tbody>
-        <% months.forEach((element) =>{ %>
-          <tr>
-            <td><%= element.month %></td>
-            <td><%= '20 ' + element.available %></td>
-            <td>
-              <button id="seeReport" class="button is-rounded is-small is-link" monthavailable="<%= element.available %>" <%if(element.availableOptions){%>disabled<%}%>>
-                <span class="icon is-small">
-                  <i class='bx bxs-file-find bx-sm'></i>
-                </span>
-                <span>Ver Reporte</span>
-              </button>
-            </td>
-            <td>
-              <button id="generatePdf" class="button is-rounded is-small is-danger" monthavailable="<%= element.available %>" <%if(element.availableOptions){%>disabled<%}%>>
-                <span class="icon is-small">
-                  <i class='bx bxs-file-pdf bx-sm'></i>
-                </span>
-                <span>Generar PDF</span>
-              </button>
-            </td>
-            <td>
-              <button id="generateReport" class="button is-rounded is-small is-primary" monthavailable="<%= element.available %>" <%if(element.availableOptions){%>disabled<%}%>>
-                <span class="icon is-small">
-                  <i class='bx bxs-file-doc bx-sm'></i>
-                </span>
-                <span>Detalle Excel</span>
-              </button>
-            </td>
-          </tr>
-        <%})%>
-      </tbody>
-    </table>
-  </div>
-  <div class="content" id="showreport"></div>
-</section>
-<script>
-  let seller;
-  <%if(!sellers){%>
-    const table = document.querySelector('#reportTable');
-    seller = '<%- seller %>';
-    removeClass(table, 'is-hidden');
-  <%}else{%>
-    seller = document.querySelector('#selectSeller').value;
-    const selectSeller = document.querySelector('#selectSeller');
-    selectSeller.onchange = changeSelect;
-  <%}%>
-
-  function changeSelect(){
-    const table = document.querySelector('#reportTable');
-    const div = document.querySelector('#showreport');
-    seller = document.querySelector('#selectSeller').value;
-    div.innerHTML = '';
-    removeClass(table, 'is-hidden');
-  }
-
-  function nameReport() {
-    let name = '1Ecommerce';
-    if (document.getElementById("selectSeller")) {
-      let index = document.getElementById("selectSeller").options.selectedIndex;
-      name = document.getElementById("selectSeller").options[index].text;
-    }
-    return name;
-  }
-
-  live('#generateReport','click', function(){
-    if (!this.disabled) {
-      let name = nameReport();
-      let month = this.getAttribute('monthavailable');
-      let loadingBlock = document.querySelector('#loading-block');
-      if(hasClass(loadingBlock,'is-hidden')){removeClass(loadingBlock,'is-hidden');}
-      io.socket.post('/generatereport', {seller, month}, (resData, jwRes) =>{
-        if(jwRes.statusCode == 200) {
-          saveAs(new Blob([resData],{type:"application/octet-stream"}), `Reporte ${name}.xlsx`);
-          addClass(loadingBlock,'is-hidden');
-        }
-      });
-    }
-  });
-
-  live('#generatePdf','click', function(){
-    if (!this.disabled) {
-      let name = nameReport();
-      let month = this.getAttribute('monthavailable');
-      let loadingBlock = document.querySelector('#loading-block');
-      if(hasClass(loadingBlock,'is-hidden')){removeClass(loadingBlock,'is-hidden');}
-      io.socket.post('/reportseller', {seller, month}, (resData, jwRes) =>{
-        if(jwRes.statusCode == 200) {
-          saveAs(new Blob([resData],{type:"application/octet-stream"}), `Factura ${name}.pdf`);
-          addClass(loadingBlock,'is-hidden');
-        }
-      });
-    }
-  });
-
-  live('#seeReport','click', function(){
-    if (!this.disabled) {
-      let month = this.getAttribute('monthavailable');
-      let loadingBlock = document.querySelector('#loading-block');
-      if(hasClass(loadingBlock,'is-hidden')){removeClass(loadingBlock,'is-hidden');}
-      window.location.href = "/showreport/" + seller + '/?month=' + month;
-    }
-  });
-</script>
