+++ views/pages/sellers/sellers.ejs
@@ -1,3 +1,4 @@
+<script src="https://cdnjs.cloudflare.com/ajax/libs/FileSaver.js/2.0.0/FileSaver.min.js"></script>
 <div class="container">
   <% if (error != undefined || error != null) { %>
     <div class="notification is-danger is-light" id="error">
@@ -23,6 +24,7 @@
           <li class="is-active"><a option="infobasics" >Información General</a></li>
           <li <%if(action!='edit'){%>class="is-hidden"<%}%>><a option="integrations">Integraciones</a></li>
           <li <%if(action!='edit'){%>class="is-hidden"<%}%>><a option="commissions">Comisiones</a></li>
+          <li <%if(action!='edit'){%>class="is-hidden"<%}%>><a option="reports">Informe de Ventas</a></li>
         </ul>
       </div>
       <div id ="infobasics">
@@ -702,6 +704,100 @@
               console.log(date);
             });
           });
+        </script>
+      </div>
+      <div class="is-hidden" id="reports">
+        <div id="loading-block" class="notification is-hidden">
+          <i class="fas fa-sync fa-spin"></i>
+            Estamos generando tu informe...
+            <progress class="progress is-small is-primary" max="100">15%</progress>
+        </div>
+        <table id="reportTable" class="table is-striped is-size-6 is-fullwidth">
+          <thead>
+            <tr>
+              <th>Mes</th>
+              <th>Fecha Disponibilidad</th>
+              <th></th>
+              <th></th>
+              <th></th>
+            </tr>
+          </thead>
+          <tbody>
+            <% months.forEach((element) =>{ %>
+              <tr>
+                <td><%= element.month %></td>
+                <td><%= '20 ' + element.available %></td>
+                <td>
+                  <button id="seeReport" class="button is-rounded is-small is-link" monthavailable="<%= element.available %>" <%if(element.availableOptions){%>disabled<%}%>>
+                    <span class="icon is-small">
+                      <i class='bx bxs-file-find bx-sm'></i>
+                    </span>
+                    <span>Ver Reporte</span>
+                  </button>
+                </td>
+                <td>
+                  <button id="generatePdf" class="button is-rounded is-small is-danger" monthavailable="<%= element.available %>" <%if(element.availableOptions){%>disabled<%}%>>
+                    <span class="icon is-small">
+                      <i class='bx bxs-file-pdf bx-sm'></i>
+                    </span>
+                    <span>Generar PDF</span>
+                  </button>
+                </td>
+                <td>
+                  <button id="generateReport" class="button is-rounded is-small is-primary" monthavailable="<%= element.available %>" <%if(element.availableOptions){%>disabled<%}%>>
+                    <span class="icon is-small">
+                      <i class='bx bxs-file-doc bx-sm'></i>
+                    </span>
+                    <span>Detalle Excel</span>
+                  </button>
+                </td>
+              </tr>
+            <%})%>
+          </tbody>
+        </table>
+        <script>
+          let seller;
+          <%if(action==='edit'){%>
+            seller = '<%- seller.id %>';
+            live('#generateReport','click', function(){
+              if (!this.disabled) {
+                let name = '<%- seller.name %>';
+                let month = this.getAttribute('monthavailable');
+                let loadingBlock = document.querySelector('#loading-block');
+                if(hasClass(loadingBlock,'is-hidden')){removeClass(loadingBlock,'is-hidden');}
+                io.socket.post('/generatereport', {seller, month}, (resData, jwRes) =>{
+                  if(jwRes.statusCode == 200) {
+                    saveAs(new Blob([resData],{type:"application/octet-stream"}), `Reporte ${name}.xlsx`);
+                    addClass(loadingBlock,'is-hidden');
+                  }
+                });
+              }
+            });
+  
+            live('#generatePdf','click', function(){
+              if (!this.disabled) {
+                let name = '<%- seller.name %>';
+                let month = this.getAttribute('monthavailable');
+                let loadingBlock = document.querySelector('#loading-block');
+                if(hasClass(loadingBlock,'is-hidden')){removeClass(loadingBlock,'is-hidden');}
+                io.socket.post('/reportseller', {seller, month}, (resData, jwRes) =>{
+                  if(jwRes.statusCode == 200) {
+                    saveAs(new Blob([resData],{type:"application/octet-stream"}), `Factura ${name}.pdf`);
+                    addClass(loadingBlock,'is-hidden');
+                  }
+                });
+              }
+            });
+  
+            live('#seeReport','click', function(){
+              if (!this.disabled) {
+                let month = this.getAttribute('monthavailable');
+                let loadingBlock = document.querySelector('#loading-block');
+                if(hasClass(loadingBlock,'is-hidden')){removeClass(loadingBlock,'is-hidden');}
+                window.location.href = "/showreport/" + seller + '/?month=' + month;
+              }
+            });
+          <%}%>
         </script>
       </div>
     </div>
