+++ views/partials/adminmenu.ejs
@@ -54,7 +54,6 @@
     </p>
     <ul class="menu-list">
       <%if(req.session.user.rights.name=='superadmin' || req.session.user.rights.permissions.includes('showsellers')){%><li><a href="/sellers">Sellers</a></li><%}%>
-      <%if(req.session.user.rights.name=='superadmin' || req.session.user.rights.permissions.includes('report')){%><li><a href="/reports">Reportes</a></li><%}%>
       <%if(req.session.user.rights.name=='superadmin' || req.session.user.rights.permissions.includes('mensajesmeli')){%><li><a href="/messagesmeli">Preguntas Mercadolibre</a></li><%}%>
     </ul>
     <%}%>
